+++ Mod.cs
@@ -1,4 +1,5 @@
 ﻿using Colossal.IO.AssetDatabase;
+using Colossal.Localization;
 using Game;
 using Game.Modding;
 using Game.SceneFlow;
@@ -8,6 +9,7 @@ namespace ExtraHotkeys
     public class Mod : IMod
     {
         public static ModSettings ModSettings;
+        private LocalizationManager LocalizationManager => GameManager.instance.localizationManager;
 
         public void OnLoad(UpdateSystem updateSystem)
         {
@@ -19,11 +21,10 @@ namespace ExtraHotkeys
             ModSettings = new ModSettings(this);
             ModSettings.RegisterInOptionsUI();
             ModSettings.RegisterKeyBindings();
-
             AssetDatabase.global.LoadSettings(nameof(ExtraHotkeys), ModSettings, new ModSettings(this));
             ModSettings.ApplyAndSave();
 
-            GameManager.instance.localizationManager.AddSource("en-US", new LocaleEN(ModSettings));
+            AddLocalizations();
 
             updateSystem.UpdateAt<UISystem>(SystemUpdatePhase.UIUpdate);
         }
@@ -37,6 +38,12 @@ namespace ExtraHotkeys
                 ModSettings.UnregisterInOptionsUI();
                 ModSettings = null;
             }
+        }
+
+        public void AddLocalizations()
+        {
+            LocalizationManager.AddSource("en-US", new LocaleEN(ModSettings));
+            // TODO: Add more languages
         }
     }
 }